requirements.txt
twelvelabs==1.0.1
pytest==8.4.2
python-dotenv==1.1.1